Make RangeVariable support floats and make all Numbers floats.
-Integers are only used for colors
-Further separate integer and number everywhere
-Add FloatSeekBar
-Modify SeekBarRangeVariableWidget to use FloatSeekBar.

Fixes #103 that has been long standing even though just filed.

+++ remixer_annotation/src/test/resources/inputs/RangeVariableMethodAnnotationTest/WrongRange.java
@@ -24,5 +24,5 @@ public class WrongRange {
    * minValue > maxValue, should fail.
    */
   @RangeVariableMethod(minValue = 15, maxValue = 0)
-  public void fail(Integer i){}
+  public void fail(Float i){}
 }
